Fix property violations due to overflow

+++ c/array-fpi/ss1.c
@@ -43,6 +43,6 @@ int main()
 		sum[0] = sum[0] + a[i];
 	}
 
-	__VERIFIER_assert(sum[0] == N*(N+2));
+	__VERIFIER_assert(sum[0] == (long long)N*(N+2));
 	return 1;
 }
